+++ packages/charts/src/components/legend/types.ts
@@ -6,7 +6,7 @@
  * Side Public License, v 1.
  */
 
-import type { LegendItemExtraValues, LegendValue } from '../../common/legend';
+import type { LegendItem, LegendItemExtraValues, LegendValue } from '../../common/legend';
 import type { SeriesIdentifier } from '../../common/series_id';
 import type {
   LegendItemListener,
@@ -34,6 +34,7 @@ export interface SharedLegendItemProps {
   labelOptions: LegendLabelOptions;
   colorPicker?: LegendColorPicker;
   action?: LegendAction;
+  legendActionOnHover: boolean;
   onClick?: LegendItemListener;
   onLegendItemMouseOver: (seriesIdentifiers: SeriesIdentifier[], path: LegendPath) => void;
   onLegendItemMouseOut: BasicListener;
@@ -43,4 +44,9 @@ export interface SharedLegendItemProps {
   toggleDeselectSeriesAction: typeof onToggleDeselectSeriesAction;
   legendTitle?: string;
   hiddenItems: number;
+}
+
+/** @internal */
+export interface LegendItemProps extends SharedLegendItemProps {
+  item: LegendItem;
 }